requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
commandlines==0.4.1
cycler==0.10.0
decorator==4.4.2
defcon==0.6.0
Deprecated==1.2.7
Flask==1.1.1
font-v==1.0.2
fontbakery==0.7.22
fonttools==4.6.0
freetype-py==2.1.0.post1
fs==2.4.11
gast==0.3.3
gftools==0.3.6
gitdb==4.0.2
GitPython==3.1.0
glyphsLib==5.1.8
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
networkx==2.4
numpy==1.18.2
opentype-sanitizer==8.0.0.post2
pandas==1.0.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.1.0
pkg-resources==0.0.0
prompt-toolkit==3.0.4
protobuf==3.11.3
ptyprocess==0.6.0
PyGithub==1.47
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
sklearn==0.0
smmap==3.0.1
soupsieve==2.0
svgpathtools==1.3.3
svgwrite==1.4
tabulate==0.8.7
tensorboard==1.14.0
tensorboardX==1.8
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
torch==1.5.0
torchvision==0.5.0
tqdm==4.43.0
traitlets==4.3.3
ttfautohint-py==0.4.3
ufoLib2==0.6.2
ufolint==1.0.0
unicodedata2==13.0.0.post2
Unidecode==1.1.1
urllib3==1.25.8
vttLib==0.9.0
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.12.1
